samples/requirements.txt
Flask==1.1.1
paho-mqtt==1.4.0